requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.18
neo4j>=5.15.0
pymupdf>=1.23.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pillow>=10.1.0
pytesseract>=0.3.10
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
tiktoken>=0.5.0
